refactor(test): split MySQL process tests
